arduino
   UART1 (D0/D1 ) Serial1 API and examples

+++ hardware/arduino/ameba/system/libameba/sw/lib/sw_lib/mbed/hal/serial_api.h
@@ -57,7 +57,7 @@ void serial_irq_handler(serial_t *obj, uart_irq_handler handler, uint32_t id);
 void serial_irq_set    (serial_t *obj, SerialIrq irq, uint32_t enable);
 
 int  serial_getc       (serial_t *obj);
-void serial_putc       (serial_t *obj, int c);
+void serial_putc       (serial_t *obj, uint8_t c);
 int  serial_readable   (serial_t *obj);
 int  serial_writable   (serial_t *obj);
 void serial_clear      (serial_t *obj);
